// src/database/types/license.ts
export interface License {
  id?: string;
  contentItemId?: string;
  applicantFirstName?: string;
  licenseUniqNumber?: string;
  applicantLastName?: string;
  licenseDescriptor?: string;
  businessName?: string;
  phoneNumber?: string;
  cellNumber?: string;
  email?: string;
  location?: string;
  additionalInfo?: string;
  parcelNumber?: string;
  quickRefNumber?: string;
  expirationDate?: string;
  isLicenseTypeStatusesOrderedList?: boolean;
  ownerName?: string;
  createdUtc?: string;
  author?: string;
  modifiedUtc?: string;
  setLicenseLocation?: string;
  setAdditionalInfo?: string;
  setParcelNumber?: string;
  setQuickRefNumber?: string;
  paymentReceived?: string;
  isAllowAssigned?: boolean;
  renewalStatusId?: boolean;
  licenseStatusId?: boolean;
  statusId: string;
  licenseStatusDisplayText?: string;
  licenseTypeDisplayText?: string;
  longitudeField: string;
  latitudeField: string;
  streetRouteField?: string;
  cityField?: string;
  stateField?: string;
  countryField?: string;
  postalCodeField?: string;
  isAllowViewInspection?: boolean;
  isAllowAddAdminNotes?: boolean;
  licenseSubType?: string;
  licenseTag?: string;
  licenseTypeName?: string | null;
  // new added
  isPaymentReceived?: string;
  isShowInspectionTab: boolean;
  displayText?: string;
  statusColor?: string;
  licenseStatus?: string;
  isManualAddress?: string;
  licenseType?: string;
  licenseTypeId?: string;
  DataType?: string;
  Id?: string;
  isForceSync?: boolean;
  LicenseTag?: string;
  SyncModel?: any;
  viewOnlyAssignUsers: boolean;
  isApiUpdateQuickRefNumberAndParcelNumber?: boolean;
  licenseNumber: string;
  isForceSyncSuccess: boolean;

  licenseStatusColor?: string;
  published: boolean;
  isEditable: boolean;
  assignedUsers?: string;
  isAllowEditLicense: boolean;
  appointmentStatus: string | null;
  appointmentStatusWithLabel: string | null;
  attachDocuments: string;
  attachedDocGridView: any;
  attachedItemCount: number;
  attachedItems: string;
  authorName: string | null;
  autoNumber: string | null;
  certFormat: string | null;
  contactEmail: string;
  contactFirstName: string;
  contactLastName: string;
  contactMailingAddress: string;
  contactPhoneNumber: string;
  contentTypes: string | null;
  dataType: string | null;
  effectiveDate: string;
  isLicense: boolean;
  issueDate: string;
  laserficheSSOUrl: string | null;
  latest: boolean;
  liabilityInsuranceExpDate: string;
  licenseOwner: string;
  licenseTypeRoles: string | null;
  licenseTypeWordDocUrl: string | null;
  mediaFileCount: number;
  nextExpirationDate: string | null;
  onlinePaymentLink: string | null;
  onlinePaymentText: string | null;
  owner: string;
  ownerAddress: string;
  ownerCellPhone: string;
  ownerEmail: string;
  ownerMailingAddress: string;
  ownerPhoneNumber: string;
  ownerUserName: string | null;
  parentContractors: string;
  publicComment: string | null;
  publicCommentCount: number;
  quickRefNumberName: string | null;
  relatedLicenseAddressCount: number;
  renewalApprovedStatusId: string | null;
  renewalStatus: string;
  returnUrl: string | null;
  showUniquePrompt: boolean;
  totalCost: string;
}